Migrate migrations to v2
