+++ .gitignore
@@ -86,7 +86,8 @@ packages/contract-wrappers/src/contract_wrappers/generated/
 packages/metacoin/src/contract_wrappers
 packages/fill-scenarios/src/generated_contract_wrappers/
 packages/order-watcher/src/generated_contract_wrappers/
-packages/migrations/src/contract_wrappers
+packages/migrations/src/v1/contract_wrappers
+packages/migrations/src/v2/contract_wrappers
 
 # solc-bin in sol-compiler
 packages/sol-compiler/solc_bin/
