+++ packages/0x.js/package.json
@@ -104,7 +104,7 @@
         "@0xproject/order-utils": "^0.0.5",
         "@0xproject/order-watcher": "^0.0.2",
         "@0xproject/sol-compiler": "^0.5.0",
-        "@0xproject/types": "^0.7.0",
+        "@0xproject/types": "0.7.0",
         "@0xproject/typescript-typings": "^0.3.2",
         "@0xproject/utils": "^0.6.2",
         "@0xproject/web3-wrapper": "^0.6.4",
