+++ packages/contract-wrappers/test/global_hooks.ts
@@ -1,5 +1,5 @@
 import { devConstants } from '@0xproject/dev-utils';
-import { runMigrationsAsync } from '@0xproject/migrations';
+import { runV1MigrationsAsync } from '@0xproject/migrations';
 import * as path from 'path';
 
 import { constants } from './utils/constants';
@@ -15,5 +15,5 @@ before('migrate contracts', async function(): Promise<void> {
         from: devConstants.TESTRPC_FIRST_ADDRESS,
     };
     const artifactsDir = `../migrations/artifacts/1.0.0`;
-    await runMigrationsAsync(provider, artifactsDir, txDefaults);
+    await runV1MigrationsAsync(provider, artifactsDir, txDefaults);
 });
