+++ packages/contracts/src/utils/formatters.ts
@@ -1,8 +1,9 @@
 import { BigNumber } from '@0xproject/utils';
+import { SignedOrder } from '@0xproject/types';
 import * as _ from 'lodash';
 
 import { orderUtils } from './order_utils';
-import { BatchCancelOrders, BatchFillOrders, MarketBuyOrders, MarketSellOrders, SignedOrder } from './types';
+import { BatchCancelOrders, BatchFillOrders, MarketBuyOrders, MarketSellOrders } from './types';
 
 export const formatters = {
     createBatchFill(signedOrders: SignedOrder[], takerAssetFillAmounts: BigNumber[] = []): BatchFillOrders {
