+++ packages/contracts/src/utils/order_utils.ts
@@ -1,10 +1,11 @@
+import { Order, SignedOrder, UnsignedOrder } from '@0xproject/types';
 import { BigNumber } from '@0xproject/utils';
 import { Web3Wrapper } from '@0xproject/web3-wrapper';
 import ethUtil = require('ethereumjs-util');
 import * as _ from 'lodash';
 
 import { crypto } from './crypto';
-import { CancelOrder, MatchOrder, OrderStruct, SignatureType, SignedOrder, UnsignedOrder } from './types';
+import { CancelOrder, MatchOrder, SignatureType } from './types';
 
 export const orderUtils = {
     createFill: (signedOrder: SignedOrder, takerAssetFillAmount?: BigNumber) => {
@@ -22,7 +23,7 @@ export const orderUtils = {
         };
         return cancel;
     },
-    getOrderStruct(signedOrder: SignedOrder): OrderStruct {
+    getOrderStruct(signedOrder: SignedOrder): Order {
         const orderStruct = {
             senderAddress: signedOrder.senderAddress,
             makerAddress: signedOrder.makerAddress,
