+++ packages/contracts/test/exchange/transactions.ts
@@ -1,4 +1,5 @@
 import { BlockchainLifecycle } from '@0xproject/dev-utils';
+import { SignedOrder, Order } from '@0xproject/types';
 import { BigNumber } from '@0xproject/utils';
 import * as chai from 'chai';
 import * as ethUtil from 'ethereumjs-util';
@@ -20,9 +21,7 @@ import {
     AssetProxyId,
     ERC20BalancesByOwner,
     ExchangeStatus,
-    OrderStruct,
     SignatureType,
-    SignedOrder,
     SignedTransaction,
 } from '../../src/utils/types';
 import { provider, txDefaults, web3Wrapper } from '../../src/utils/web3_wrapper';
@@ -47,7 +46,7 @@ describe('Exchange transactions', () => {
     let erc20Balances: ERC20BalancesByOwner;
     let signedOrder: SignedOrder;
     let signedTx: SignedTransaction;
-    let order: OrderStruct;
+    let order: Order;
     let orderFactory: OrderFactory;
     let makerTransactionFactory: TransactionFactory;
     let takerTransactionFactory: TransactionFactory;
