+++ packages/fill-scenarios/package.json
@@ -38,7 +38,7 @@
     "dependencies": {
         "@0xproject/base-contract": "^0.3.2",
         "@0xproject/order-utils": "^0.0.5",
-        "@0xproject/types": "^0.7.0",
+        "@0xproject/types": "0.7.0",
         "@0xproject/typescript-typings": "^0.3.2",
         "@0xproject/utils": "^0.6.2",
         "@0xproject/web3-wrapper": "^0.6.4",
