+++ packages/order-watcher/package.json
@@ -82,7 +82,7 @@
         "@0xproject/fill-scenarios": "^0.0.2",
         "@0xproject/json-schemas": "0.7.22",
         "@0xproject/order-utils": "^0.0.5",
-        "@0xproject/types": "^0.7.0",
+        "@0xproject/types": "0.7.0",
         "@0xproject/typescript-typings": "^0.3.2",
         "@0xproject/utils": "^0.6.2",
         "@0xproject/web3-wrapper": "^0.6.4",
